+++ src/Landing/__tests__/index.test.js
@@ -1,0 +1,10 @@
+import Landing from ".."
+
+describe("Landing component", () => {
+  const wrapper = shallow(<Landing/>)
+
+  it("renders correctly", () => {
+    expect(wrapper).toMatchSnapshot()
+  })
+
+})