+++ src/TestStoreContext/index.js
@@ -1,18 +1,18 @@
 import React from 'react'
 import {withStore} from '../db'
 import initValues from '../db/initialValues.json'
+import {Button, TextField, Typography} from '@material-ui/core'
 
 
 export const TestStoreContext = ({store}) =>
   <div>
-    <h3>Testing store context</h3>
-    <button onClick={() => store.handleHelloWorld(store.value)}>Click me!</button>
-    <input autoFocus onChange={e => store.handleChangeValue(e.target.value)} value={store.value}/>
+    <Typography variant="h3">Testing store context</Typography>
+    <Button onClick={() => store.handleHelloWorld(store.value)}>Click me!</Button>
+    <TextField onChange={e => store.handleChangeValue(e.target.value)} value={store.value}/>
     {
       initValues.value !== store.value &&
-      <button onClick={() => store.handleChangeValue("")}>Clear</button>
+      <Button onClick={() => store.handleChangeValue("")}>Clear</Button>
     }
-    <hr/>
   </div>
 
 export default withStore(TestStoreContext)