+++ src/hooks/useDimensions.js
@@ -1,0 +1,23 @@
+import {useState, useEffect} from "react"
+
+/**
+ * Hook that returns the correct window size,
+ * even after resize
+ */
+export default function useDimensions() {
+
+  const [width, setWidth] = useState(window.innerWidth)
+  const [height, setHeight] = useState(window.innerHeight)
+
+  const handleResize = () => {
+    setWidth(window.innerWidth)
+    setHeight(window.innerHeight)
+  }
+
+  useEffect(() => {
+    window.addEventListener('resize', handleResize)
+    return () => window.removeEventListener('resize', handleResize)
+  }, [])
+
+  return ({width, height})
+}