+++ src/lib/router.js
@@ -1,0 +1,5 @@
+export const routes = {
+  ROOT: "/",
+  TEST: "/test",
+  FORGOT_PASSWORD: "/new-password"
+}